refactor: unify SimulationSummaryContent and SimulationSummaryExpanded
The two components were ~90% identical (476 vs 475 lines): same
modifier/compute/fix/variable tables, same columns, same handlers.
They had already drifted — the expanded view was missing the 'show
simulation box' row present in the overlay view.

SimulationSummaryContent now accepts optional expanded/onShowLess
props controlling only the outer wrapper chrome (minimize button,
wrapper class); SimulationSummaryExpanded becomes a 19-line wrapper
around it. Restores the 'show simulation box' row to the expanded
view. No behavior change to the overlay/modal callers.
